backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pydantic==1.10.13
email-validator==2.1.0
